Refactor exam session (removed all service and controllers from manage service). Moved exceptions to conduct
