+++ exam-conduct-service/src/main/java/com/sebrown2023/controller/ApiErrorDto.java
