+++ exam-conduct-service/src/main/java/com/sebrown2023/service/ExamService.java
@@ -1,7 +1,7 @@
 package com.sebrown2023.service;
 
 import com.sebrown2023.model.db.Exam;
-import com.sebrown2023.model.exceptions.ExamSessionException;
+import com.sebrown2023.exceptions.ExamSessionException;
 
 import java.util.UUID;
